+++ R/slide.R
@@ -33,6 +33,7 @@ replace_fn_names <- function(fn, replace = list()){
 #' of the window. If `.size` is even for center alignment, "centre-right" & "centre-left"
 #' is needed.
 #' @param .bind If `.x` is a list, should `.x` be combined before applying `.f`?
+#' If `.x` is a list of data frames, row binding is carried out.
 #'
 #' @rdname slide
 #' @export
@@ -44,6 +45,9 @@ replace_fn_names <- function(fn, replace = list()){
 #' the purrr-like syntax. For some specialist functions like `mean` and `sum`,
 #' you may like to check out for
 #' [RcppRoll](https://CRAN.R-project.org/package=RcppRoll) for faster performance.
+#'
+#' `slide()` is intended to work with list (and column-wise data frame). To
+#' perform row-wise sliding window on data frame, please check out [pslide()].
 #'
 #' @examples
 #' x <- 1:5
@@ -134,6 +138,28 @@ slide_dfc <- function(
 #' slide2(df, df, ~ ., .size = 2)
 #' pslide(lst, ~ ., .size = 1)
 #' pslide(list(lst, lst), ~ ., .size = 2)
+#'
+#' ###
+#' # row-wise sliding over data frame
+#' ###
+#'
+#' my_df <- data.frame(
+#'   group = rep(letters[1:2], each = 8),
+#'   x = c(1:8, 8:1),
+#'   y = 2 * c(1:8, 8:1) + rnorm(16),
+#'   date = rep(as.Date("2016-06-01") + 0:7, 2)
+#' )
+#' 
+#' slope <- function(...) {
+#'   data <- list(...)
+#'   fm <- lm(y ~ x, data = data)
+#'   coef(fm)[[2]]
+#' }
+#' 
+#' my_df %>% 
+#'   nest(-group) %>% 
+#'   mutate(slope = purrr::map(data, ~ pslide_dbl(., slope, .size = 2))) %>% 
+#'   unnest()
 slide2 <- function(
   .x, .y, .f, ..., .size = 1, .fill = NA, .partial = FALSE, 
   .align = "right", .bind = FALSE
@@ -288,9 +314,7 @@ slider <- function(
   .align = "right", .bind = FALSE
 ) {
   bad_window_function(.size)
-  if (!is.list(.x) && .bind) {
-    abort(sprintf("`.bind = TRUE` only accepts list, not %s.", typeof(.x)))
-  }
+  abort_not_lst(.x, .bind = .bind)
   if (is.data.frame(.x)) .x <- as.list(.x)
   len_x <- NROW(.x)
   abs_size <- abs(.size)
@@ -314,7 +338,7 @@ slider <- function(
   lst_idx <- seq_len(len_x - abs_size + 1)
   if (sign < 0) lst_idx <- rev(lst_idx) + 1
   out <- purrr::map(lst_idx, function(idx) .x[idx:(idx + sign * (abs_size - 1))])
-  if (.bind) return(bind_lst(out)) else out
+  if (.bind) bind_lst(out) else out
 }
 
 #' @rdname slider
@@ -331,13 +355,11 @@ pslider <- function(
 
 bind_lst <- function(x) {
   type_elements <- flatten_chr(purrr::modify_depth(x, 2, ~ typeof(.)[1]))
-  type <- purrr::reduce(type_elements, identical_type)
-  if (type == "list") {
-    out <- lapply(x, dplyr::bind_rows)
+  if (all(type_elements == "list")) {
+    lapply(x, dplyr::bind_rows)
   } else {
-    out <- lapply(x, function(y) do.call(c, y))
+    lapply(x, dplyr::combine)
   }
-  out
 }
 
 bad_window_function <- function(.size) {
@@ -380,15 +402,16 @@ pad_slide <- function(x, .size = 1, .fill = NA, .align = "right") {
   if (is_null(.fill)) return(x) 
   fill_size <- abs(.size) - 1
   if (align == "right") {
-    return(c(rep(.fill, fill_size), x))
+    c(rep(.fill, fill_size), x)
   } else if (align %in% c("c", "center", "centre", "cl", "center-left", "centre-left")) {
     lsize <- floor(fill_size / 2)
-    return(c(rep(.fill, lsize), x, rep(.fill, fill_size - lsize)))
+    c(rep(.fill, lsize), x, rep(.fill, fill_size - lsize))
   } else if (align %in% c("cr", "center-right", "centre-right")) {
     lsize <- ceiling(fill_size / 2)
-    return(c(rep(.fill, lsize), x, rep(.fill, fill_size - lsize)))
+    c(rep(.fill, lsize), x, rep(.fill, fill_size - lsize))
+  } else {
+    c(x, rep(.fill, fill_size)) # "left"
   }
-  c(x, rep(.fill, fill_size)) # "left"
 }
 
 bind_df <- function(x, .size, .fill = NA, .id = NULL, byrow = TRUE) {
@@ -399,9 +422,11 @@ bind_df <- function(x, .size, .fill = NA, .id = NULL, byrow = TRUE) {
   }
   lst <- new_list_along(x[[abs_size]])
   lst[] <- .fill
-  if (byrow) 
-    return(dplyr::bind_rows(lst, !!! x[-seq_len(abs_size - 1)], .id = .id))
-  dplyr::bind_cols(lst, !!! x[-seq_len(abs_size - 1)])
+  if (byrow) {
+    dplyr::bind_rows(lst, !!! x[-seq_len(abs_size - 1)], .id = .id)
+  } else {
+    dplyr::bind_cols(lst, !!! x[-seq_len(abs_size - 1)])
+  }
 }
 
 check_valid_window <- function(.size, .align) {
@@ -410,5 +435,11 @@ check_valid_window <- function(.size, .align) {
       "Can't use `.align = %s` for even window `.size`. Please use `.align = 'center-left'` or `.align = 'center-right'`.",
       .align
     ))
+  }
+}
+
+abort_not_lst <- function (.x, .bind = FALSE) {
+  if (!is.list(.x) && .bind) {
+    abort(sprintf("`.bind = TRUE` only accepts list, not %s.", typeof(.x)))
   }
 }