+++ R/tidyr-verbs.R
@@ -13,7 +13,7 @@
 #' stocks %>% gather(stock, price, -time)
 gather.tbl_ts <- function(data, key = "key", value = "value", ...,
   na.rm = FALSE, convert = FALSE, factor_key = FALSE) {
-  key <- enexpr(key)
+  key <- sym(enexpr(key))
   new_key <- c(key(data), key)
   value <- enexpr(value)
   quos <- enquos(...)
@@ -57,7 +57,7 @@ spread.tbl_ts <- function(data, key, value, fill = NA, convert = FALSE,
     abort(sprintf("`key` must not be `%s`, as it's the `index`.", key_var))
   }
   key_left <- setdiff(key_vars(data), key_var)
-  new_key <- key(key_reduce(data, .vars = key_left, validate = FALSE))
+  new_key <- key(key_remove(data, .vars = key_left, validate = FALSE))
 
   tbl <- spread(
     as_tibble(data), key = !! key, value = !! value, fill = fill, 
@@ -138,29 +138,26 @@ unnest.lst_ts <- function(data, ..., key = id(),
     list_cols <- setdiff(list_cols, preserve)
     quos <- syms(list_cols)
   }
-  if (length(quos) == 0) {
-    return(data)
-  }
+  if (length(quos) == 0) return(data)
+
   nested <- transmute(ungroup(data), !!! quos)
 
   # checking if the nested columns has `tbl_ts` class (only for the first row)
   first_nested <- slice(nested, 1)
   eval_df <- purrr::imap(first_nested, dplyr::first)
   is_tsbl <- purrr::map_lgl(eval_df, is_tsibble)
-  if (is_false(any(is_tsbl))) {
-    return(NextMethod())
-  }
+  if (is_false(any(is_tsbl))) return(NextMethod())
+
   if (sum(is_tsbl) > 1) {
     abort("Only accepts a list-column of `tbl_ts` to be unnested.")
   }
   out <- as_tibble(data) %>% 
     unnest(!!! quos, .drop = .drop, .id = .id, .sep = .sep, .preserve = .preserve)
-  tsbl <- eval_df[[is_tsbl]]
+  tsbl <- eval_df[is_tsbl][[1L]]
   idx <- index(tsbl)
   validate <- FALSE
-  if (is_empty(key)) {
-    validate <- TRUE
-  }
+  if (is_empty(key)) validate <- TRUE
+
   key <- c(key(tsbl), key)
   idx_chr <- quo_text(idx)
   # restore the index class, as it's dropped by NextMethod()
@@ -224,9 +221,10 @@ semi_join.lst_ts <- anti_join.lst_ts
 as_lst_ts <- function(x) {
   grped_df <- dplyr::is_grouped_df(x)
   if (grped_df) {
-    return(structure(
+    structure(
       x, class = c("lst_ts", "grouped_df", "tbl_df", "tbl", "data.frame")
-    ))
+    )
+  } else {
+    structure(x, class = c("lst_ts", "tbl_df", "tbl", "data.frame"))
   }
-  structure(x, class = c("lst_ts", "tbl_df", "tbl", "data.frame"))
 }